+++ src/utils/index.js
@@ -17,3 +17,5 @@ export * from './checkboxList';
 export * from './getFileExtension';
 export * from './generateSettingsLabel';
 export * from './stringToWords';
+export * from './getDropdownOptions';
+export * from './fetchJsonShemas';