+++ test/bigtest/mocks/index.js
@@ -6,3 +6,4 @@ export * from './associated-mapping-profile';
 export * from './associated-action-profiles';
 export * from './form-config-samples';
 export * from './job-profiles-child-wrappers';
+export * from './json-schemas';